+++ packages/analytics/analytics-chart/src/enums/chart-metric-display.enum.ts
@@ -1,6 +1,0 @@
-
-export enum ChartMetricDisplay {
-  Hidden = 'hidden',
-  SingleMetric = 'single',
-  Full = 'full',
-}