+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.6.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.14...@kong-ui-public/analytics-metric-provider@8.6.15) (2025-01-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [8.6.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@8.6.13...@kong-ui-public/analytics-metric-provider@8.6.14) (2025-01-13)